test(model): outsource getAngles function into test:algebra.h
'tis makes using it outside of the original file possible
