+++ test/model/algebra/algebra.cpp
@@ -3,6 +3,6 @@
 #include <catch2/catch.hpp>
 #include <vector>
 
-TEST_CASE("") {
-
+TEST_CASE("Points and Vectors can be used interchangeably") {
+    
 }
